refactor: mapper구조 변경
