+++ src/main/java/org/scoula/mapper/feedback/AIAnalysisReportMapper.java
@@ -1,4 +1,4 @@
-package org.scoula.mapper;
+package org.scoula.mapper.feedback;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
