+++ src/main/java/org/scoula/mapper/learning/LearningMapper.java
@@ -1,4 +1,4 @@
-package org.scoula.mapper;
+package org.scoula.mapper.learning;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
