+++ src/main/java/org/scoula/mapper/trading/TradingMapper.java
@@ -1,4 +1,4 @@
-package org.scoula.mapper;
+package org.scoula.mapper.trading;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
