+++ src/main/java/org/scoula/service/Auth/UserService.java
@@ -1,10 +1,7 @@
 package org.scoula.service.Auth;
 
-import org.apache.ibatis.annotations.Param;
 import org.scoula.domain.Auth.vo.UserVo;
 import org.scoula.domain.type.dto.RiskTypeDto;
-import org.scoula.mapper.UserMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
