+++ src/main/java/org/scoula/service/Auth/UserServiceImpl.java
@@ -1,9 +1,8 @@
 package org.scoula.service.Auth;
 
-import org.apache.ibatis.annotations.Param;
 import org.scoula.domain.Auth.vo.UserVo;
 import org.scoula.domain.type.dto.RiskTypeDto;
-import org.scoula.mapper.UserMapper;
+import org.scoula.mapper.auth.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
