+++ src/main/java/org/scoula/service/journal/InvestmentJournalServiceImpl.java
@@ -2,7 +2,7 @@ package org.scoula.service.journal;
 
 import org.scoula.domain.journal.dto.InvestmentJournalDto;
 import org.scoula.domain.journal.vo.InvestmentJournalVO;
-import org.scoula.mapper.InvestmentJournalMapper;
+import org.scoula.mapper.journal.InvestmentJournalMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
