+++ src/main/java/org/scoula/service/learning/LearningContentServiceImpl.java
@@ -1,8 +1,7 @@
 package org.scoula.service.learning;
 
-import lombok.RequiredArgsConstructor;
 import org.scoula.domain.learning.dto.ContentDto;
-import org.scoula.mapper.LearningContentMapper;
+import org.scoula.mapper.learning.LearningContentMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
