+++ src/main/java/org/scoula/service/learning/LearningGptAsyncHelper.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.scoula.domain.learning.dto.GptLearningContentResponseDto;
 import org.scoula.domain.learning.vo.LearningContentVO;
 import org.scoula.domain.learning.vo.LearningQuizVO;
-import org.scoula.mapper.LearningMapper;
+import org.scoula.mapper.learning.LearningMapper;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
