+++ src/main/java/org/scoula/service/learning/LearningSaveService.java
@@ -3,7 +3,7 @@ package org.scoula.service.learning;
 import lombok.RequiredArgsConstructor;
 import org.scoula.domain.learning.vo.LearningContentVO;
 import org.scoula.domain.learning.vo.LearningQuizVO;
-import org.scoula.mapper.LearningMapper;
+import org.scoula.mapper.learning.LearningMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
