+++ src/main/resources/org/scoula/mapper/auth/UserMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.scoula.mapper.UserMapper">
+<mapper namespace="org.scoula.mapper.auth.UserMapper">
 
     <update id="updateRiskType" parameterType="map">
         UPDATE user
