+++ src/main/resources/org/scoula/mapper/feedback/JournalFeedbackMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.scoula.mapper.JournalFeedbackMapper">
+<mapper namespace="org.scoula.mapper.feedback.JournalFeedbackMapper">
 
     <resultMap id="journalResultMap" type="org.scoula.domain.journal.vo.InvestmentJournalVO">
         <id property="id" column="id"/>
