+++ src/main/resources/org/scoula/mapper/learning/LearningContentMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.scoula.mapper.LearningContentMapper">
+<mapper namespace="org.scoula.mapper.learning.LearningContentMapper">
 
     <select id="selectByRiskType" parameterType="string" resultType="org.scoula.domain.learning.dto.ContentDto">
         SELECT
