+++ src/main/resources/org/scoula/mapper/mocktrading/HoldingMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.scoula.mapper.HoldingMapper">
+<mapper namespace="org.scoula.mapper.trading.HoldingMapper">
 
     <!-- ResultMap 정의 -->
     <resultMap id="holdingResultMap" type="org.scoula.domain.mocktrading.vo.Holding">
