+++ src/main/resources/org/scoula/mapper/trading/StockMapper.xml
